+++ benchmark/Streamly/Benchmark/Data/Stream/StreamK.hs
@@ -31,8 +31,8 @@ import Prelude hiding
 import qualified Prelude as P
 import qualified Data.List as List
 
+import qualified Streamly.Internal.Control.Concurrent as S
 import qualified Streamly.Internal.Data.Stream.StreamK as S
-import qualified Streamly.Internal.Data.SVar as S
 
 import Gauge (bench, nfIO, bgroup, Benchmark, defaultMain)
