+++ src/Streamly/Internal/Data/Producer.hs
@@ -46,7 +46,7 @@ where
 
 import Streamly.Internal.Data.Stream.StreamD.Step (Step(..))
 import Streamly.Internal.Data.Stream.StreamD.Type (Stream(..))
-import Streamly.Internal.Data.SVar (defState)
+import Streamly.Internal.Data.SVar.Type (defState)
 import Streamly.Internal.Data.Unfold.Type (Unfold(..))
 
 import Streamly.Internal.Data.Producer.Type
