+++ src/Streamly/Internal/Data/Stream/IsStream/Transform.hs
@@ -237,11 +237,12 @@ import Streamly.Internal.Data.Stream.IsStream.Common
     , interjectSuffix
     , intersperseM
     )
+import Streamly.Internal.Control.Concurrent (MonadAsync)
 import Streamly.Internal.Data.Stream.Prelude (fromStreamS, toStreamS)
 import Streamly.Internal.Data.Stream.Serial (SerialT)
 import Streamly.Internal.Data.Stream.StreamD (fromStreamD, toStreamD)
 import Streamly.Internal.Data.Stream.StreamK (IsStream)
-import Streamly.Internal.Data.SVar (MonadAsync, Rate(..))
+import Streamly.Internal.Data.SVar (Rate(..))
 import Streamly.Internal.Data.Time.Units (TimeUnit64, AbsTime, RelTime64)
 
 import qualified Streamly.Internal.Data.Fold as FL
