+++ src/Streamly/Internal/Data/Stream/Zip.hs
@@ -51,10 +51,10 @@ import GHC.Exts (IsList(..), IsString(..))
 import Text.Read
        ( Lexeme(Ident), lexP, parens, prec, readPrec, readListPrec
        , readListPrecDefault)
+import Streamly.Internal.Control.Concurrent (MonadAsync)
 import Streamly.Internal.BaseCompat ((#.), errorWithoutStackTrace, oneShot)
 import Streamly.Internal.Data.Stream.StreamK (IsStream(..), Stream)
 import Streamly.Internal.Data.Maybe.Strict (Maybe'(..), toMaybe)
-import Streamly.Internal.Data.SVar (MonadAsync)
 
 import qualified Streamly.Internal.Data.Stream.Parallel as Par
 import qualified Streamly.Internal.Data.Stream.Prelude as P
